requirements.txt
openai>=1.30.0
pydantic>=2.0
networkx>=3.0
streamlit>=1.35.0
plotly>=5.20.0
pyvis>=0.3.2
rich>=13.0
numpy>=1.26.0
python-dotenv>=1.0.0
tavily-python>=0.5.0
